人気メニューを集計して調べられるようにする #72
人気メニュー確認機能
