+++ resources/views/home.blade.php
@@ -25,6 +25,8 @@
                     @can('browsing')
                     <a href="/sum_total" class="btn btn-primary my-1">売上確認</a>
                     <a href="/orders" class="btn btn-primary my-1">注文・お会計確認</a>
+                    <a href="/aggregate" class="btn btn-primary my-1">人気商品確認</a>
+                    
                     <div class="form-group col-auto">
                         <label for="seat">席を選択してメニュー表示</label>
                         {{ Form::select('seat', \App\Seat::forSelect(), old('seat'), empty($errors->first('seat')) ? ['class'=>"form-control", 'id'=>'seat'] : ['class'=>"form-control is-invalid", 'id'=>'seat']) }}
