+++ resources/views/payments/sum_total.blade.php
